Finished ScalaBufSerializer, MessageBufSerializer
ScalaBufSerializer currently cant write the snapshot because in the
currently older version there exists a limit on the size ProtoBuf can
read from a stream to prevent malicious attacks. We are forced to use
this version because eventsourced would need to recompile their
templates otherwise.

MessageBufSerializer is the new alternative using msgbuf protocol. The
de-/serialization process is more dynamic but with the @Index annotation
it should be flexible enough for foreward and backward compatibility.